Do not allow passphrases to be typed with non-ASCII keymaps (#681250)
Keymaps are complicated, and it's impossible to know whether a
particular passphrase will be typable at boot time, but we can be pretty
sure that anything from a non-ASCII keymap will not be typable. If the
user is attempting to use such a keymap in the passphrase dialog,
disable input and display an error message.