Change "failed to download" messages from critical to warning.
This message gets logged if a single mirror fails, in which case DNF
moves on to another.  That's not critical, and the presence of that log
level is causing kickstart_tests to assume a test failed.  We also still
have an exception that gets raised should DNF run out of mirrors which
handles the really critical case.