Add a script to print an IP address for the host.
Python's server address determination depends on the hostname, which is
not particularly reliable, so grep through ip addr output to find an
address that is actually configured. It doesn't matter which one, since
the virt guest will have the host as the default route, and, barring any
weird firewall configurations, once the host gets the packet it will
know that it's local.