Make sure anaconda reads in ks file from OEMDRV device. (#1057271)
Even if dracut detects an OEMDRV labelled device and mounts it etc.,
the installer itself does not know there is a ks file to read in since
ks= or inst.ks= was never set on the commandline.

So, instead check explicitly for the existence of ks.cfg where it
gets mounted in such a situation and then set the appropriate flags
to ensure installation actually proceeds automatically.

This should finish off the series of disjoint OEMDRV patches..

Resolves: rhbz#1057271