Ignore requests to create existing groups (#1261002)
Part of the fallout from removing libuser is that some errors that were
lost somewhere across the fork and chroot are now actually errors. One
of those errors is that you are not allowed to request that a group be
created if that group already exists, so it is written. Unfortunately
the current code to add users to existing groups depends on requests to
create existing groups not crashing, so go back to ignoring the error as
a quick fix for 23.