Fix potential issues with the username guesser.
Add a new signal handler that can be blocked during username guessing
instead of cleaning up variables after the fact and hoping nothing was
asynchronous.

Also quit using a dict as a boolean, remove the references to the
hostname guessing that never existed that was probably the reason for
the dict, and remove gender-specific language.