Use signal handlers in the user spoke more sensibly.
Use different signal handlers for different widgets in the advanced user
dialog so that unrelated portions of the UI do not need to be checked
for every change. Remove an unnecessary call to the advanced user button
signal handler that is handled by the initial state of the dialog. Use
the arguments passed in to the signal handlers.