Remove unnecessary grab_focus and set_sensitive calls
Two of the conditionals around the grab_focus calls (password checkbox
set but no password, full name set but username not set) are only
possible via kickstart in which case the UI should be left alone anyway,
and the default case is already handled by focusWidgetName. The
sensitivity of the advanced button is handled by the input signal
handlers and does not need to be doubled in refresh().