Update the password strength bar during the password strength check.
There's no need to validate the password twice, and it would make more
sense anyway for the strength reported by the strength bar to match the
strength we're using to enforce policy.

Also, one of the checks was bypassing the input delay that I went
through all the trouble of adding. Checking the password on every
keystroke can case UI freezes and we should avoid doing it.

Initialize the strength bar to "0" in glade so that entering the spoke
does not display a medium strength before the initial check runs.