Update the on-disk snapshot of storage when adv. disks are added (#1267944)
If users add advanced storage devices we need to recreate the on-disk snapshot
of storage because it's really a new on-disk state. Also, adding such devices is
a change in disk selection so we need to revert to the snapshot first before
going there.

(cherry picked from commit 4d4ac4fea6c979b2d028ec1273f2e7233d731fbd)