dracut: accept inst.dd=[file:]/dd.iso (#1268792)
Like with inst.ks=/ks.cfg, some users want to be able to inject the
driver disk into the initramfs and load it from there.

The mechanisms for doing this are already basically present, we just
need to handle it the same way we do when the kickstart is already
inside the initramfs: check to see if the requested file exists, and if
it's already present, use it as-is.

The canonical form of this argument is:

    inst.dd=file:/path/to/dd.iso

`dd.iso` can be an ISO or a driver RPM, which must be injected into the
initramfs.
`file:` may be omitted, although that practice is discouraged.
`path:` can be used instead of `file:`, but that's also discouraged.

Resolves: rhbz#1268792