Split exception description from exception traceback
When traceback will be part of the exception message split the
description from traceback. Description is used in name of the
bug in Bugzilla.
This is useful when saving exception in exception handler and
raising this exception elsewhere (subprocess exception).

This is based on patch from BlivetGUI. Thank you Vojtech Trefny.