By default, skip those kickstart tests we know to be failing.
This adds a new knownfailure test type.  I've put it on both of the
tests that always fail, and added +x back on the multilib test.  The
permission bit was always something I thought of as more of a local
one-off way to disable a test anyway.

Note you can still run the failing tests with "-t knownfailure", if you
want to see if they are still failing.  If you want to test just one,
you can also pass it on the command line.