decode package name for /etc/sysconfig/kernel (RHBZ #1261569)
Seems in Python 3 when we get a package name from RPM in this
way it comes out as a bytestring, we need a regular string as
we're going to write it into a text file. Without this fix
/etc/sysconfig/kernel gets a line:

DEFAULTKERNEL=b'kernel-core'

which obviously isn't what we wanted, and means new kernels
sometimes don't get set as the default boot option.

Signed-off-by: Samantha N. Bueno <sbueno+anaconda@redhat.com>