Store fewer kinds of things in the dirinstall option.
The behavior we want is that if --dirinstall is specified without
arguments use a default path, if --dirinstall is specified with
arguments use that argument as the path, and if --dirinstall is
specified with no arguments do not use a path. So it's a good thing that
argparse that can do that on its own.