Store fewer kinds of things in the dirinstall option.
The behavior we want is that if --dirinstall is specified without
arguments use a default path, if --dirinstall is specified with
arguments use that argument as the path, and if --dirinstall is not
specified do not use a path. Argparse can do that on its own, so let it
do so instead of doing strange things with booleans.