Changes in network state revalidate sources rhbz#1270354
When altering the network state, it may result in selected
network sources becoming inaccessable due to typos in DNS
or GATEWAY.

With this patch, after the network spoke completes with
changes to the network config, it triggers the software
payload to revalidate its state.

If the state becomes invalid, the source spoke errors will guide
the user towards what is wrong.

Resolves: rhbz#1270354