Improve install space required estimation (#1224048)
The problem was that 35% add is not enough on the smaller installations.
That result in crash on RPM transaction check. RPM required more space
than we have.
This problem was consulted with RPM team and this is result.
The RPM rounding size of files to fragment size (default 4KiB) for
better estimation also they are required package header size * 4 for RPM
DB.

We don't have header size or file size in the installation so I'm trying
to simulate worst case scenario for file rounding and using 2KiB for
every file to get something like RPM DB space required.
Also 10% is add to all this as a safeguard.

Seems as stable assumption right now.
+
Improved logging for packaging.

Resolves: rhbz#1224048