rpmostreepayload: Fix two issues with mounting
This makes the internal mounts show up a lot more cleanly
in the output of `mount` at least after installation.

This might fix:
https://bugzilla.redhat.com/show_bug.cgi?id=1274977
as it's possible Blivet/systemd will be less confused.

Basically I believe that becuase later mounts were traversing the
`/mnt/sysroot` -> `$deploy/sysroot` bind mount, they were being
`canonicalized` to point into the target.  This messed up the
hierarchy.

That's one bug - the other silly bug is that we were missing a `-o
remount` option for the `/usr` bind mount, causing it to be mounted
twice.