rpmostreepayload: Also unmount internal mounts during shutdown
I got a report that using livemedia-creator, the sysroot mount
was hanging around.  We previously on the RHEL7.2 branch in
https://github.com/cgwalters/anaconda/commit/31c0ebc5016724bcb6bdb948f969fea62d01f5ce
added a special method for the payload to clean up mounts,
but it turns out in at least this f23 branch there is already done
in `Payload.unsetup()`.

Use that method to unmount everything we mounted internally during
shutdown.