Make reclaim work with small screens and big labels
Allow the reclaim window to scroll horizontally since, just as a for
instance, sometimes the label we're trying to display got set to "Red
Hat Enterprise Linux Server 7.2 for ppc64le" which does not fit in the
tiny windows that the system I just described insists upon using. Also
allow the disk label column to be shrunk in case the user doesn't care
about all of that.