Improve language selection at low resolutions.
Move the two language strings (native and English) into a single column
so that the strings in each row can be positioned independent of other
rows. Allow the English string to be ellipsized so that the really long
ones (e.g., Norwegian Bokmal) don't cut off the right side of the
TreeView and mess up the layout.

Allow the locale selection pane to scroll horizontally in order to
handle long locale names (one of the Chinese locales does not fit at
800x600), and change the column sizing from Grow Only to Automatic to
allow the column to shrink back when switching to other locales. This
way the scrollbar is only shown when needed.