Fix dasdfmt progress bar. (#1210450)
As a means of performance improvement, dasdfmt now defaults to formatting 10
cylinders in one step instead of 01. However, this causes issues with updating
the progress bar when dasdfmt is run. So, make a slight adjustment in our
calculation of % formatted to accurately take into account the new dasdfmt
changes. Previously we'd always add one to completedCylinders -- one per step.
But now add 10 to completedCylinders -- 10 for each step.
For more details on this dasdfmt feature, see bz1053828.

Resolves: rhbz#1210450