Add window title (#1280077)
Even though the window header is usually hidden during normal installation
and the title text can't be seen, there are still some cases where empty window
title looks weird - such as when using Anaconda for dirinstall or image install,
on some live spins, etc.

So set the title to "Anaconda Installer" to handle these cases and also
make sure to retranslate it if user changes the language on the Welcome
spoke.