Prevent duplicate kernel params being written in zipl.conf (#1029505)
If upgrading and choosing to create a new boot loader config, anaconda
always wrote duplicate kernel parameters in zipl.conf. This was because
on the zipl screen, when prompting the user to add additional boot loader
args, the entry box was already populated with a set of pre-created args,
the user tacked on anymore that they needed, and anaconda appended that
whole line. It should really only be appending additional args.

So, make the additional boot loader args entry box empty from the start,
until/unless the user adds any of their own.

Resolves: rhbz#1029505