Make sure python3.5 code can run in early initrd
There are 2 problems with running python3.5 early in the initrd when
entropy is low and the nonblocking pool hasn't been initialized yet:

1. cpython runs getrandom() syscall to see the object id hash. This
blocks until the pool is initialized and isn't critical from a security
standpoint. Override this with PYTHONHASHSEED=42

2. Python3.5 uses getrandom() syscall for urandom(), this also blocks
instead of falling back to reading /dev/urandom like it did in py3.4
this is fixed by monkeypatching os.urandom to just read /dev/urandom
in parse-kickstart -- important because we still want unique UUIDs for
networking.