Automatically generate sr@latin.po
The modern Serbian language exhibits synchronic digraphia, which is an
interesting few minutes on Wikipedia and means that it can be written in
two different ways. The Serbian translators provide a translation in
Cyrillic, and from that we can generate the translation in Latin. Stop
pulling sr@latin.po from Zanata, since no one updates it there, and add
a new Makefile rule to generate one from sr.po using a filter program
that comes with gettext.