Make it possible to skip saving of kickstarts and logs (#1285519)
Make it possible to skip saving of kickstart (both input & output)
and logs to the installed system by adding the save_kickstarts and
save_logs CLI-/boot-options.

The main use case is image generation that might use internal
repository URLs and other internal resources that would then
otherwise be included in kickstart and/or logs on the resulting
image. Other scenarios, such as OEM OS installs, are also likely to
benefit from this added functionality.

Resolves: rhbz#1285519