Add possibility to have /usr and /var parts on iscsi (#856446)
Prevents adding of _netdev fstab parameter for mountpoints which are
on the same LUN where / mountpoint is.
This will fix /usr and /var are not mounted on boot and /boot is not
mounted in singleuser mode.

We support /var and /usr on a LUN only when the / is on the same LUN.

Resolves: rhbz#856446