Refactor variable name (#914637)
Refactored code to eliminate a generic variable name. Changed 'tuple' to
be device_data as a more accurate description of what it contains. This
is a precursor to another patch where tuple is used on a wider basis.

Added four missing constants that define fields in the device_data
variable. Found cases in the file where the index number had been used
and replaced with these new constants.

Related: rhbz#914637