Addon repos should still be added to anaconda-ks.cfg.
Before, we were leaving out addon repos on media installs because the
file paths didn't make sense - it wasn't anything the user set up,
so there was no way for them to know what baseurl to give.

However, we were still writing out groups to the %packages section that
were only a part of that addon.  Thus, the kickstart file was broken.

The fix here is to write the addon repos to anaconda-ks.cfg, but give
them a special baseurl of file://addon.  anaconda will interpret this
when it reads the file in and look for a valid addon for the base repo
by that name.

Resolves: rhbz#1099178