Fix multipath UI representation (#914637)
Changed the depopulate method of the FilterWindow class to correctly
identify and remove outdated representations of a multipath device in
the user interface. Added a check to ensure the first member device of a
multipath device is not accidentally displayed and counted as both an
mpath device member and on the tab for other san devices.

The issues impacting the display of multipath devices in the user
interface were also contributing to an improper count of selected and
total devices in the user interface. Added a data structure to track
tallied storage devices and account for the movement of multipath
devices from single other san devices over to the multipath tab.

Related: rhbz#1014425
Resolves: rhbz#914637