Add specification for the user interaction config file
The specification aims to provide a way for Anaconda and secondary
system setup tools to communicate during the OS installation phase.

With a way to communicate the installation process can be streamlined
so that configuration screens are not needlessly shown to the user
multiple times. It also makes it possible to run tools before
Anaconda starts up (eq. a language sleection tool in a live environment)
and to provide initial configuration file snippets to disable screen in
Anaconda
or in other system setup tools.