Add a finished method to spokes (#1300499)
This method can be used to clean up a spoke before starting the
installation. The method is called when continue is clicked on the
summary hub. It is completely optional and if it doesn't exist on the
spoke it will be skipped.

Add finished method to the Network spoke to handle killing nmce and
disconnecting callbacks so that they don't cause problems if the
installation triggers network changes.