Make sure devices configured by NM from iBFT have correct ifcfg (#831002)
Related: rhbz#831002

Because of possible race in loader (media install, setting lang and keyboard
not fast enough) for devices activated by NM with iBFT connection ifcfg file
might not be created with BOOTPROTO=ibft value.  It is required for
compatibility and supplying dracut with ip=ibft option at the end of
installation.  Also handles case of multiple ibft devices.