Don't delete shared/boot parts in deleteAll (#1183880)
This approach changes the deleteAll behaviour as follows:

* Partitions known to be shared with other OSes won't be deleted
* Boot partitions won't be deleted if there are unknown parts

You can delete these partitions only by specifically targeting
them. It also adds a variant of the ConfirmDeleteDialog for
boot partitions, with text specifically explaining that they
might be needed for other OSes to boot.

Based on the PR from AdamWill. In order to apply this PR it was reworked
becuase multiselection feature changed structure of the custom spoke a
lot.

This should be final fix for the UEFI shared partitions problem.

Related: rhbz#1183880