Do not skip evaluation after removing partitions
The on_remove_clicked() method in Custom spoke used return instead
of break. That skipped the evaluation after the for cycle with
multiselection.

Problem happens when user selects multiple mountpoints. Confirm some and
then skip others. This could set the Custom spoke to undefined behavior
because the post code is not evaluated.