Don't remove selected shared part when Delete all (#1183880)
When user click on any not-shared partition and check option Delete all
it will remove all not-shared partitions from that page.
If an user did this on shared partition it will remove all not-shared
partitions in that page but it will also remove that selected
partition (even when it's shared).

To stay consistent use the same rule for the selected partition.

Related: rhbz#1183880