Use CSS to style the internal widgets.
Provide names for the stylable internal widgets, and use CSS instead of
Pango attributes and markup to set the style attributes. This allows the
appearance of the anaconda widgets to be more easily modified by
product-specific stylesheets.

The new widget names are only applied to the leaf widgets (labels and
buttons and such) since making the containers mess-with-able sounds like
trouble, and the type and existence of the containers could change if we
decide to switch to gadget types.

As part of the changes for moving things to CSS, what looks like the
intent of anaconda_mountpoint_selector_toggle_background (unselected
selectors are gray, selected selectors are black) actually works now,
and the incomplete state of AnacondaSpokeSelector is now tied to the
inconsistent widget flag.

The CSS data is stored in regular CSS files that get compiled into
libAnacondaWidgets.so, since that seemed less gross than a bunch of
multi-line C strings.

(cherry picked from commit c25430fc39eaf0349a51d975201b210c31e91162)