Change the default locale to C.UTF-8 (#1312607)
C.UTF-8 is a newish glibc locale this is supposed to solve the problem
of how to get an untranslated locale that supports multi-byte
characters. I mean, it's still English, but not so overtly. And it's
also the only locale besides regular old 7-bit C that's guaranteed to be
installed anymore.

(cherry picked from commit 949d4454b45d2e6fa3508d297abdfb6630fdceb5)