Hotfix for missing storage in payload class (#1271657)
The problem here is that payload is setting up storage when
PayloadManager run a thread but the user could do steps which
will call the spaceRequired() method sooner.

Now the behavior is the same as it was with the old spaceRequired
method for the variant when storage is not set.

Resolves: rhbz#1271657
(cherry picked from commit 4b87177e277ecfbcd2c760e27dc91c357449dec8)