Do not change the GUI language to a missing locale. (#1312607)
If the user attempts to change the language to a locale that is not
available, fallback to the default language instead. This will prevent a
translation from being used where the localization data for the
translation, such as the text direction, is missing.

(cherry picked from commit 39dcb536e50fc872e1b40fe88757d8cc402b2577)