don't install kernel-PAE on x86_64 (#1313957)
PAE is *available* on x86_64 systems, but we do not want to
install kernel-PAE on them even if it is in the repo for some
reason.

An alternative to this is to rename `isPaeAvailable()` to
`isPaeWanted()` and have it return False on x86_64, but I guess
that's technically an interface change, and this is also easy.

Signed-off-by: Vratislav Podzimek <vpodzime@redhat.com>