Run actions for argparse arguments (#1285519)
Previously we would just set the value to the argument destination,
but that meant custom Action classes would never be used and even
built-in actions might not behave as expected as they would actually
never be called.

So instead just call the action, which should fix these issues.

Related: rhbz#1285519