Fix create device with bad name when parsing KS (#1293539)
When installation is started with these parameters:

ks=file:/ks.cfg ifname=provisioning:52:54:00:48:8b:3f
bootdev=provisioning ip=provisioning:dhcp

and in kickstart file you have this line:

network --device 52:54:00:48:8b:3f --bootproto dhcp --noipv6

Then the kickstart file is parsed and ksdev0 is created but the device
should be named as provisioning because of the cmdline command
ifname=provisioning:52:54:00:48:8b:3f.
This caused fail in the dracut because it's waiting on ksdev0 and
provisioning devices which are physically only one device.

Resolves: rhbz#1293539