Create ibft vlan ifcfg file in stage2 if needed (#831002)
Resolves rhbz#831002

In some cases, eg

- DVD install without requiring network in stage1
- NeworkManager winning race when configuring network from ibft for
  kickstart

ifcfg file for vlan connection was not created and therefore ip=ibft was not
added to dracut options.

This is update of commit 5f97dfde6b5ce18c7362ba03719f21d8b6f6df5a,
now handling also the vlan case actually.

Note that the ifcfg file will be ignored by NM in installer (having
BOOTPROTO=ibft) so its creating later in stage2 should be safe. We care about
the file because of the dracut parameter setting and for compatibility.