Clean up devicetree filtering of md containers.
Correctly apply the container code only to container devices,
and not also to the member disks which also return a value for
udev_device_get_md_container.

We don't allow installation to degraded md, so there's no point
in allowing containers with any ignored members.

Lastly, we have to take into account both the ignored and
exclusive disk lists when deciding if member disks are to be
ignored or not.

Related: rhbz#1269872
Related: rhbz#1120640