Add screen access manager
The screen access manager is based on the user interaction config spec[0].
It tracks which screens have the users visited and possibly also which
options on the screen have the users changed.

The results are then written out to the /etc/sysconfig/anaconda file
so that post-installation setup tools know which screens (and possibly
settings) have been seen/changed by the user, so that they don't need
to show them again.

Anaconda also parses any existing user interaction config files. This
makes it possible to react to any pre-installation setup tools, such as
a language selection tool. If any screen is marked as visited in an
existing user interaction config file Anaconda will not show it during
the installation.

[0] see doscs/user-interaction-config-file-spec.rst