network: disable NM for parent device of vlan configured in ibft (#831002)
Resolves: rhbz#831002

We set NM_CONTROLLED to "no" for devices set up for iSCSI in initramfs (boot
from iSCSI, root on iSCSI). If the device is vlan we need to set it also for
its parent.

If the parent device of vlan device configured in iBFT is controlled by NM
it causes I/O errors when shutting down due to too early device takedown.
It can also cause iSCSI target disconnection during boot when NM is started
if network service is on (which is by default after installation).
If the network service is off, it still disconnects iSCSI target if the
parent device is activated (ONBOOT=yes).