Make windows in metacity closable (#1319590)
Display a close button on non-anaconda windows (yelp and
nm-connection-editor). This way the windows can be closed gracefully
without needing to use keyboard shortcuts. This will not affect anaconda
windows, since no titlebar is shown on anaconda windows.

In order to prevent external programs from being maximized, thus hiding
the titlebar and the close button, add more config overrides to disable
maxizing the window when the titlebar is double-clicked and when the
window is dragged to the edge of the screen.

Thanks to Ray Strode (halfline) for the button-layout change.