Require that the English locale data be available. (#1315494)
For en_US.UTF-8 to work as a default locale, it needs to be defined on
the system, which is not guaranteed in this exciting modern world of
ours. Ridiculously, glibc provides two distinct ways of actually
installing the data for a particular locale, so look for either of them.