Don't use booleans in Requires (#1323314)
We're not allowed to use booleans in Requires because dnf-by-default did
not include the tools used to construct a Fedora compose. This means
that boot.iso and other situations will end up with both
glibc-langpack-en and glibc-all-langpacks, but that's apparently the
world we live in now.

(cherry picked from commit 4405ad4315eae9cde872da96904ba07e5ceb8b52)