Fix problem of DASDs not being formatted in text ks. (#1259437)
Unformatted and/or LDL DASDs were not properly selected if doing a
text-based kickstart installation. This fixes those problems in a manner
much more in line with the GUI now.

run_dasdfmt() now takes to_format (list of DASDs to format) as an optional
kwarg, since you can't specify this if doing a ks install (too early on in the
storage thread), but you must specify it if you're doing a manual installation
(otherwise run_dasdfmt always runs).

Resolves: rhbz#1259437