Add support for kickstart %onerror scripts (#1412538)
These run on most error conditions (including when %traceback scripts
would run), but note that they do not run on other kickstart script
failures. That would be strange.

The runOnErrorScripts function is in iutil, not kickstart, because it
gets imported all over the place and this would lead to import cycles.

(cherry picked from commit e470d17)

Resolves: rhbz#1412538