Refresh metadata when updates checkbox changes (#1211907)
When the state of a repo is changed the metadata needs to be refreshed.
To complicate matters updateBaseRepo rereads the repos from disk,
resetting the state of updates and updates-testing to enabled.

This patch adds a method to dnfpayload to control whether or not to
include updates. This is used right after the reload to set the correct
state on the repos so that the user selection won't be reset by the
upadate.

(cherry picked from commit f40cbe4f520e672721b612927489b84d82f3605d)