Ignore some E1101 no-member errors when running pylint
Some of these are due to a lack of introspection of the module from
pylint (eg. BlockDev) and the GLib.GError one is possibly a pylint bug
since it is supposed to work with GLib.

For now we ignore tham as specifically as possible so that real E1101
errors will be easy to find.